Move Metadata object out of KafkaConsumerActor for better Java APIs
